Update symbolic link for ghost git

+++ docker/containers/ghost/ghost_cmd.sh
@@ -9,11 +9,15 @@ sudo chown -R ghost:ghost $GHOST_USER_DIR
 
 npm rebuild # For different environments
 
-if [! -d $DOCKER_ROOT_DIR ]; then
+if [ ! -d $DOCKER_ROOT_DIR ]; then
     sudo mkdir -p $DOCKER_ROOT_DIR
 fi
 
-sudo ln -s $GHOST_ROOT_DIR/.git $DOCKER_ROOT_DIR/.git
+# Symbolically link git folder for bower and submodule resolution
+if [ ! -L $DOCKER_ROOT_DIR/.git ]; then
+    sudo ln -s $GHOST_ROOT_DIR/.git $DOCKER_ROOT_DIR/.git
+fi
+
 # rm -rf .git && git init && git submodule update --init --recursive # reset submodule root path
 
 npm install -g grunt-cli knex-migrator
